Problem 5 Initial Commit and Solution
I solved problem 5, by first determining an initial product that would
be divisible by all numbers in the range of 1-20, with a slight initial
optimization which would exclude factors which are factors of other
numbers in the range. So, for example, since 2 is a factor of 4, and 4
is a factor of 8, and 8 is a factor of 16, then 2, 4, and 8 would not be
included in the initial product, since by including 16 in the initial
product, we are implicitly including 2, 4, and 8. Then, once that
initial product is made, there is still a chance that, through the
consecutive multiplications, some factors were included redundantly in
the product as a side effect of the multiplications of the various
factors. So, a post processing loop goes through all of the factors, and
for each one, tries to extract the factor from the product (as long as
the range of 1-20 is still evenly divisible into the resulting product).
It keeps extracting instances of a factor until it may no longer remove
any more, and then it moves onto the next factor, repeating the process
until the factor of 1 is reached (which should be skipped, since it may
infinitely be extracted). At this point, we have the minimal product
which is evenly divisible by the range 1-20.